requirements.txt
# Requirements for 3D model generation
torch>=2.0.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.25.0
trimesh>=4.0.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0

# Required by LGM-full model
xformers>=0.0.22
kiui>=0.2.0
einops>=0.7.0

